fix loading a fursuit as performer
A fursuit can have another fursuit listed as its performer. This is used
when the second fursuit represents someone's main fursona, and the first
is an additional character they portray.

When a entity like this was encountered, the program was returning null
as the performer, while the intended fallback is to use the name specified.

The intended logic in resolvePerformer checks if the result of
loadPerformer is falsy, then falls back to loadCharacter. However,
loadPerformer is an async function, and resolvePerformer wasn't awaiting
on it; thus it was checking whether the returned Promise is falsy, which
is never true.

This commit fixes loading fursuits as performers by awaiting on
loadPerformer, thus only falling back if the unwrapped value is falsy.

fixes GH-1.