Add comments
The best code comments should not only explain what the code does or how it
achieves its goal, but give context on why the code does what it does.

Because while the former can be deduced by analyzing the code, admit it,
nobody is going to remember the latter when coming back to a code base
after a few years. This is even worse for hobby projects.