Reverse character and event name in title
Existing code have the event name come first, similar to:
"Event name - Character, character"

However, when there is not enough space available, such as in a
browser tab or when viewing an album, the title will be cut off:
"Event n..."

When viewing an album, the event name is already part of the context,
so move the character name to the front:
"Character and Character at Event Name"
When the name is cut off part of the character's name is still visible:
"Charact..."
A format that is also suitable for posting to social media.

Additionally, the format changes depending on the number of characters.
No characters: (an empty array, not a falsy value, is required)
"Event Name"
A single character:
"Character at Event Name"
Two characters:
"Character and Character at Event Name"
Three or more characters:
"Character, Character, and Character at Event Name"